Delete testing functions (for UI testing only)
