+++ src/scripts/services/authService.ts
@@ -27,10 +27,29 @@ export async function registerUser({
   return response.json();
 }
 
-export async function getUsers() {
-  const response = await fetch(
-    "https://musical-notes-backend.onrender.com/users"
-  );
-  if (!response.ok) throw new Error("Error loading users");
-  return response.json();
+export async function checkLogIn(
+  userName: string,
+  password: string
+): Promise<boolean> {
+  try {
+    const response = await fetch(
+      "https://musical-notes-backend.onrender.com/login",
+      {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ username: userName, password }),
+      }
+    );
+    if (response.ok) {
+      return true;
+    } else {
+      const data = await response.json();
+
+      console.error("Login failed:", data.message);
+      return false;
+    }
+  } catch (error) {
+    console.error("Error occurred during login:", error);
+    return false;
+  }
 }
